Add a new LtsTimeStepper base class
This allows compile-time enforcement of using LTS steppers in the LTS
code.
