+++ tests/Unit/Time/Actions/Test_UpdateU.cpp
@@ -24,6 +24,8 @@
 #include "Utilities/TaggedTuple.hpp"
 #include "tests/Unit/ActionTesting.hpp"
 
+class TimeStepper;
+
 namespace {
 struct Var : db::SimpleTag {
   static std::string name() noexcept { return "Var"; }
@@ -44,7 +46,8 @@ struct component {
   using metavariables = Metavariables;
   using chare_type = ActionTesting::MockArrayChare;
   using array_index = int;
-  using const_global_cache_tag_list = tmpl::list<OptionTags::TimeStepper>;
+  using const_global_cache_tag_list =
+      tmpl::list<OptionTags::TypedTimeStepper<TimeStepper>>;
   using action_list = tmpl::list<Actions::UpdateU>;
   using simple_tags =
       db::AddSimpleTags<Tags::TimeStep, variables_tag, history_tag>;
