+++ tests/Unit/Time/TimeSteppers/CMakeLists.txt
@@ -5,4 +5,5 @@ set(LIBRARY_SOURCES
   ${LIBRARY_SOURCES}
   TimeSteppers/Test_AdamsBashforthN.cpp
   TimeSteppers/Test_RungeKutta3.cpp
+  TimeSteppers/TimeStepperTestUtils.cpp
   PARENT_SCOPE)
