+++ tests/Unit/Time/TimeSteppers/Test_RungeKutta3.cpp
@@ -4,10 +4,6 @@
 #include "tests/Unit/TestingFramework.hpp"
 
 #include "Parallel/PupStlCpp11.hpp"
-#include "Time/History.hpp"
-#include "Time/Slab.hpp"
-#include "Time/Time.hpp"
-#include "Time/TimeId.hpp"
 #include "Time/TimeSteppers/RungeKutta3.hpp"
 #include "Time/TimeSteppers/TimeStepper.hpp"
 #include "tests/Unit/TestCreation.hpp"
@@ -18,13 +14,6 @@ SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3", "[Unit][Time]") {
   const TimeSteppers::RungeKutta3 stepper{};
   TimeStepperTestUtils::check_substep_properties(stepper);
   TimeStepperTestUtils::integrate_test(stepper, 0, 1., 1e-9);
-
-  const Slab slab(0., 1.);
-  const TimeId time_id(true, 0, slab.start());
-  TimeSteppers::History<double, double> history;
-  CHECK_FALSE(stepper.can_change_step_size(time_id, history));
-  history.insert(slab.start(), 0., 0.);
-  CHECK_FALSE(stepper.can_change_step_size(time_id, history));
 }
 
 SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Variable",
@@ -37,13 +26,6 @@ SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Backwards",
                   "[Unit][Time]") {
   const TimeSteppers::RungeKutta3 stepper{};
   TimeStepperTestUtils::integrate_test(stepper, 0, -1., 1e-9);
-
-  const Slab slab(0., 1.);
-  const TimeId time_id(false, 0, slab.end());
-  TimeSteppers::History<double, double> history;
-  CHECK_FALSE(stepper.can_change_step_size(time_id, history));
-  history.insert(slab.start(), 0., 0.);
-  CHECK_FALSE(stepper.can_change_step_size(time_id, history));
 }
 
 SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Stability",
@@ -58,18 +40,6 @@ SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Factory",
   // The Unit.Time.TimeSteppers.RungeKutta3.Factory does not need to
   // check anything
   CHECK(true);
-}
-
-SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Boundary.Equal",
-                  "[Unit][Time]") {
-  TimeStepperTestUtils::equal_rate_boundary(
-      TimeSteppers::RungeKutta3{}, 0, 1e-9, true);
-}
-
-SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Boundary.Equal.Backwards",
-                  "[Unit][Time]") {
-  TimeStepperTestUtils::equal_rate_boundary(
-      TimeSteppers::RungeKutta3{}, 0, 1e-9, false);
 }
 
 SPECTRE_TEST_CASE("Unit.Time.TimeSteppers.RungeKutta3.Serialization",
